lib/python/requirements.txt
llama-index==0.9.48
google-generativeai==0.3.2
python-dotenv==1.0.0 